IoC Refactor (AutoFac)

+++ Roton.Interface/Audio/Composition/AudioComposer.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Roton.Core;
+using Roton.Emulation.Data;
 
 namespace Roton.Interface.Audio.Composition
 {
@@ -38,7 +38,7 @@ namespace Roton.Interface.Audio.Composition
 
             _frequencyDutyCycleTable =
                 Enumerable.Range(0, 12 * 6)
-                .Select(i => outputSampleRate * AccumulatorMultiplier / (440d * Math.Pow(2d, (double)(i - 33) / 12)))
+                .Select(i => outputSampleRate * AccumulatorMultiplier / (440d * Math.Pow(2d, (i - 33d) / 12)))
                 .Select(i => (int)i)
                 .ToArray();
         }
